Improve performance of scripts/coalesce_json.py.

+++ scripts/coalesce_json.py
@@ -15,6 +15,7 @@ import uuid
 import pprint
 import itertools
 from pathlib import Path
+from collections import defaultdict, Counter
 
 from settings import output_file_path
 from pipeline.util import CromObjectMerger
@@ -29,34 +30,46 @@ seen = {}
 read = reader.Reader()
 coalesce_count = 0
 print(f'Coalescing JSON files in {path} ...')
+counter = Counter()
+files_by_id = defaultdict(list)
 for filename in files:
-	with open(filename, 'r') as fh:
-		content = fh.read()
-		try:
-			m = read.read(content)
-			id = m.id
-			if id in seen:
-				canon_file = seen[id]
-# 				print(f'*** {id} already seen in {canon_file} ; merging {filename}')
-				merger = CromObjectMerger()
-				with open(canon_file, 'r') as cfh:
-					canon_content = cfh.read()
-					n = read.read(canon_content)
-					try:
-						merger.merge(m, n)
-					except model.DataError:
-						print(f'Exception caught while merging data from {newfile}:')
-						print(d)
-						print(content)
-						raise
-				merged_data = factory.toString(m, False)
-				d = json.loads(merged_data)
-				with open(canon_file, 'w') as data_file:
-					json.dump(d, data_file, indent=2, ensure_ascii=False)
-					os.remove(filename)
-				coalesce_count += 1
-			else:
-				seen[id] = filename
-		except model.DataError as e:
-			print(f'*** Failed to read CRM data from {filename}: {e}')
+	p = Path(filename)
+	id = p.name
+	counter[id] += 1
+	files_by_id[id].append(p)
+
+for id in sorted(counter):
+	count = counter[id]
+	if count > 1:
+		files = files_by_id[id]
+		for filename in files:
+			with open(filename, 'r') as fh:
+				content = fh.read()
+				try:
+					m = read.read(content)
+					id = m.id
+					if id in seen:
+						canon_file = seen[id]
+		# 				print(f'*** {id} already seen in {canon_file} ; merging {filename}')
+						merger = CromObjectMerger()
+						with open(canon_file, 'r') as cfh:
+							canon_content = cfh.read()
+							n = read.read(canon_content)
+							try:
+								merger.merge(m, n)
+							except model.DataError:
+								print(f'Exception caught while merging data from {newfile}:')
+								print(d)
+								print(content)
+								raise
+						merged_data = factory.toString(m, False)
+						d = json.loads(merged_data)
+						with open(canon_file, 'w') as data_file:
+							json.dump(d, data_file, indent=2, ensure_ascii=False)
+							os.remove(filename)
+						coalesce_count += 1
+					else:
+						seen[id] = filename
+				except model.DataError as e:
+					print(f'*** Failed to read CRM data from {filename}: {e}')
 print(f'Coalesced {coalesce_count} JSON files in {path}')
